+++ code/components/SmartphoneScreen.js
