+++ code/tracemob/Pourquoi.js
@@ -1,189 +1,4 @@
-import { Route, Switch } from 'react-router'
-import CGU from './tracemob/CGU'
-import SmartphoneScreen from './SmartphoneScreen'
-import Meta from './Meta'
-import archives from './getTracemobArchiveArticles'
-import ReactMarkdown from 'react-markdown'
-import Emoji from 'Components/Emoji'
-
 export default () => (
-	<Switch>
-		<Route path="/tracemob/cgu">
-			<CGU />
-		</Route>
-		<Route path="/">
-			<Home />
-		</Route>
-	</Switch>
-)
-
-const Home = () => (
-	<div>
-		<div
-			css={`
-				height: 90vh;
-				box-shadow: 0 12px 12px -10px rgba(48, 41, 41, 0.1);
-				background: #375c9dff;
-				display: flex;
-				flex-direction: column;
-				align-items: center;
-				overflow: hidden;
-			`}
-		>
-			<Meta
-				title="Tracemob"
-				description="Collectez vos données de déplacement"
-			/>
-			<div
-				css={`
-					color: white;
-					width: 80%;
-					height: 70%;
-
-					text-shadow: 0 1px 4px rgba(0, 0, 0, 0.3);
-					margin: 0 auto;
-					padding-left: 1rem;
-					display: flex;
-					align-items: center;
-					header,
-					#phone {
-						width: 50%;
-					}
-					h1 {
-						font-size: 350%;
-						font-weight: 600;
-						margin: 1.2rem 0;
-						color: white;
-					}
-					p {
-						font-size: 150%;
-					}
-					@media (max-width: 800px) {
-						flex-wrap: wrap;
-						header,
-						#phone {
-							width: 100%;
-						}
-					}
-
-					@media (max-width: 800px) {
-						width: 100%;
-						h1 {
-							font-size: 250%;
-							margin: 0.4rem 0;
-						}
-						#phone {
-							width: 30%;
-						}
-						header img {
-							width: 100px;
-						}
-
-						display: flex;
-						flex-wrap: nowrap;
-						overflow: hidden;
-					}
-				`}
-			>
-				<header>
-					<img
-						src={require('./tracemob/tracemob.svg').default}
-						css="width: 200px"
-					/>
-					<h1>Tracemob</h1>
-					<p>
-						Collectez <strong>vos</strong> données <br />
-						de déplacement.
-					</p>
-				</header>
-				<SmartphoneScreen />
-			</div>
-			<div
-				css={`
-					img {
-						height: 40px;
-					}
-
-					margin: 0 auto;
-					color: white;
-				`}
-			>
-				<p>Téléchargez l'application</p>
-
-				<a href="https://apps.apple.com/us/app/tracemob/id1545371062">
-					<img src={require('./tracemob/appstore-badge.png').default} />
-				</a>
-				<a href="https://play.google.com/store/apps/details?id=com.fabmob.tracemob">
-					<img src={require('./tracemob/playstore-badge.png').default} />
-				</a>
-			</div>
-
-			<div
-				href="#quoi"
-				css={`
-					width: 100%;
-					a {
-						display: block;
-						background: white;
-						border-radius: 0.3rem;
-						width: 14rem;
-						text-align: center;
-						cursor: pointer;
-						text-decoration: none;
-						padding: 0.2rem 0.6rem;
-						font-size: 130%;
-						color: black;
-						text-transform: uppercase;
-						font-weight: 600;
-						margin: 3rem auto;
-					}
-				`}
-			>
-				<a href="#quoi">😲 Comment ça ?</a>
-			</div>
-		</div>
-		<Pourquoi />
-		<Archives />
-	</div>
-)
-
-const Archives = () => (
-	<div css="background: #0000000a">
-		<div
-			css={`
-				border-top: 9px solid #375c9d;
-				box-shadow: 2px 5px 12px 3px rgba(48, 41, 41, 0.34);
-			`}
-		></div>
-		<div
-			css={`
-				margin: 0 auto;
-				max-width: 40rem;
-				padding: 0.3rem 1rem;
-			`}
-		>
-			<h2>
-				{' '}
-				<Emoji emoji="🕰️" /> Historique du projet
-			</h2>
-			<ul>
-				{archives.map((el) => (
-					<li>
-						<details>
-							<summary css="text-transform: capitalize">
-								{el.attributes.title}
-							</summary>
-
-							<ReactMarkdown source={el.body} escapeHtml={false} />
-						</details>
-					</li>
-				))}
-			</ul>
-		</div>
-	</div>
-)
-
-let Pourquoi = () => (
 	<div
 		id="quoi"
 		css={`
@@ -230,7 +45,7 @@ let Pourquoi = () => (
 		</p>
 		<img
 			className="illustration"
-			src={require('./tracemob/tracemob.svg').default}
+			src={require('./tracemob.svg').default}
 			css="width: 60px"
 		/>
 		<p>
